Bist hisselerinin geçmiş verilerinin toplanmasına yarayan endpoint hazırlandı.

+++ BorsaAgent.API/Infrastructure/Database/Models/StockDailyData.cs
@@ -13,7 +13,7 @@ public class StockDailyData
     public decimal High { get; set; }
     public decimal Low { get; set; }
     public decimal Close { get; set; }
-    public long Volume { get; set; }
+    public decimal Volume { get; set; }
     public DateTime CreatedAt { get; set; }
 
     // Navigation
